Fetch available templates from gitignore.io

This return list of all templates supported by gitignore.io.

gi_available_templates()

Value

A character with all templates supported by gitignore.io.

Details

The returned list is returned as lower case characters.

Examples

gi_available_templates()

#>   [1] "1c"                          "1c-bitrix"                  
#>   [3] "a-frame"                     "actionscript"               
#>   [5] "ada"                         "adobe"                      
#>   [7] "advancedinstaller"           "adventuregamestudio"        
#>   [9] "agda"                        "al"                         
#>  [11] "alteraquartusii"             "altium"                     
#>  [13] "amplify"                     "android"                    
#>  [15] "androidstudio"               "angular"                    
#>  [17] "anjuta"                      "ansible"                    
#>  [19] "apachecordova"               "apachehadoop"               
#>  [21] "appbuilder"                  "appceleratortitanium"       
#>  [23] "appcode"                     "appcode+all"                
#>  [25] "appcode+iml"                 "appengine"                  
#>  [27] "aptanastudio"                "arcanist"                   
#>  [29] "archive"                     "archives"                   
#>  [31] "archlinuxpackages"           "aspnetcore"                 
#>  [33] "assembler"                   "ate"                        
#>  [35] "atmelstudio"                 "ats"                        
#>  [37] "audio"                       "automationstudio"           
#>  [39] "autotools"                   "autotools+strict"           
#>  [41] "awr"                         "azurefunctions"             
#>  [43] "backup"                      "ballerina"                  
#>  [45] "basercms"                    "basic"                      
#>  [47] "batch"                       "bazaar"                     
#>  [49] "bazel"                       "bitrise"                    
#>  [51] "bitrix"                      "bittorrent"                 
#>  [53] "blackbox"                    "bloop"                      
#>  [55] "bluej"                       "bookdown"                   
#>  [57] "bower"                       "bricxcc"                    
#>  [59] "buck"                        "c"                          
#>  [61] "c++"                         "cake"                       
#>  [63] "cakephp"                     "cakephp2"                   
#>  [65] "cakephp3"                    "calabash"                   
#>  [67] "carthage"                    "certificates"               
#>  [69] "ceylon"                      "cfwheels"                   
#>  [71] "chefcookbook"                "chocolatey"                 
#>  [73] "clean"                       "clion"                      
#>  [75] "clion+all"                   "clion+iml"                  
#>  [77] "clojure"                     "cloud9"                     
#>  [79] "cmake"                       "cocoapods"                  
#>  [81] "cocos2dx"                    "cocoscreator"               
#>  [83] "code"                        "code-java"                  
#>  [85] "codeblocks"                  "codecomposerstudio"         
#>  [87] "codeigniter"                 "codeio"                     
#>  [89] "codekit"                     "codesniffer"                
#>  [91] "coffeescript"                "commonlisp"                 
#>  [93] "compodoc"                    "composer"                   
#>  [95] "compressed"                  "compressedarchive"          
#>  [97] "compression"                 "conan"                      
#>  [99] "concrete5"                   "coq"                        
#> [101] "cordova"                     "craftcms"                   
#> [103] "crashlytics"                 "crbasic"                    
#> [105] "crossbar"                    "crystal"                    
#> [107] "cs-cart"                     "csharp"                     
#> [109] "cuda"                        "cvs"                        
#> [111] "cypressio"                   "d"                          
#> [113] "dart"                        "darteditor"                 
#> [115] "data"                        "database"                   
#> [117] "datarecovery"                "dbeaver"                    
#> [119] "defold"                      "delphi"                     
#> [121] "dframe"                      "diff"                       
#> [123] "direnv"                      "diskimage"                  
#> [125] "django"                      "dm"                         
#> [127] "docfx"                       "docpress"                   
#> [129] "docz"                        "dotenv"                     
#> [131] "dotfilessh"                  "dotnetcore"                 
#> [133] "dotsettings"                 "dreamweaver"                
#> [135] "dropbox"                     "drupal"                     
#> [137] "drupal7"                     "drupal8"                    
#> [139] "e2studio"                    "eagle"                      
#> [141] "easybook"                    "eclipse"                    
#> [143] "eiffelstudio"                "elasticbeanstalk"           
#> [145] "elisp"                       "elixir"                     
#> [147] "elm"                         "emacs"                      
#> [149] "ember"                       "ensime"                     
#> [151] "episerver"                   "erlang"                     
#> [153] "espresso"                    "executable"                 
#> [155] "exercism"                    "expressionengine"           
#> [157] "extjs"                       "fancy"                      
#> [159] "fastlane"                    "finale"                     
#> [161] "firebase"                    "flashbuilder"               
#> [163] "flask"                       "flex"                       
#> [165] "flexbuilder"                 "floobits"                   
#> [167] "flutter"                     "font"                       
#> [169] "fontforge"                   "forcedotcom"                
#> [171] "forgegradle"                 "fortran"                    
#> [173] "freepascal"                  "fsharp"                     
#> [175] "fuelphp"                     "fusetools"                  
#> [177] "games"                       "gcov"                       
#> [179] "genero4gl"                   "geth"                       
#> [181] "ggts"                        "gis"                        
#> [183] "git"                         "gitbook"                    
#> [185] "go"                          "godot"                      
#> [187] "goodsync"                    "gpg"                        
#> [189] "gradle"                      "grails"                     
#> [191] "greenfoot"                   "groovy"                     
#> [193] "grunt"                       "gwt"                        
#> [195] "haskell"                     "helm"                       
#> [197] "hexo"                        "hol"                        
#> [199] "homeassistant"               "homebrew"                   
#> [201] "hsp"                         "hugo"                       
#> [203] "hyperledgercomposer"         "iar"                        
#> [205] "iar_ewarm"                   "iarembeddedworkbench"       
#> [207] "idapro"                      "idris"                      
#> [209] "igorpro"                     "images"                     
#> [211] "infer"                       "inforcms"                   
#> [213] "inforcrm"                    "intellij"                   
#> [215] "intellij+all"                "intellij+iml"               
#> [217] "ionic3"                      "jabref"                     
#> [219] "java"                        "java-web"                   
#> [221] "jboss"                       "jboss-4-2-3-ga"             
#> [223] "jboss-6-x"                   "jboss4"                     
#> [225] "jboss6"                      "jdeveloper"                 
#> [227] "jekyll"                      "jenv"                       
#> [229] "jetbrains"                   "jetbrains+all"              
#> [231] "jetbrains+iml"               "jgiven"                     
#> [233] "jigsaw"                      "jmeter"                     
#> [235] "joe"                         "joomla"                     
#> [237] "jspm"                        "julia"                      
#> [239] "jupyternotebooks"            "justcode"                   
#> [241] "kate"                        "kdevelop4"                  
#> [243] "kdiff3"                      "keil"                       
#> [245] "kentico"                     "kicad"                      
#> [247] "kirby2"                      "kobalt"                     
#> [249] "kohana"                      "komodoedit"                 
#> [251] "konyvisualizer"              "kotlin"                     
#> [253] "labview"                     "labviewnxg"                 
#> [255] "lamp"                        "laravel"                    
#> [257] "latex"                       "lazarus"                    
#> [259] "leiningen"                   "lemonstand"                 
#> [261] "less"                        "liberosoc"                  
#> [263] "librarian-chef"              "libreoffice"                
#> [265] "lilypond"                    "linux"                      
#> [267] "lithium"                     "localstack"                 
#> [269] "logtalk"                     "lsspice"                    
#> [271] "ltspice"                     "lua"                        
#> [273] "lyx"                         "m2e"                        
#> [275] "macos"                       "magento"                    
#> [277] "magento1"                    "magento2"                   
#> [279] "magic-xpa"                   "matlab"                     
#> [281] "maven"                       "mavensmate"                 
#> [283] "mdbook"                      "mean"                       
#> [285] "mercurial"                   "mercury"                    
#> [287] "meson"                       "metals"                     
#> [289] "metaprogrammingsystem"       "meteor"                     
#> [291] "meteorjs"                    "microsoftoffice"            
#> [293] "mikroc"                      "moban"                      
#> [295] "modelsim"                    "modx"                       
#> [297] "momentics"                   "monodevelop"                
#> [299] "mplabx"                      "mule"                       
#> [301] "nanoc"                       "nativescript"               
#> [303] "ncrunch"                     "nesc"                       
#> [305] "netbeans"                    "nette"                      
#> [307] "nikola"                      "nim"                        
#> [309] "ninja"                       "node"                       
#> [311] "nodechakratimetraveldebug"   "nohup"                      
#> [313] "notepadpp"                   "nova"                       
#> [315] "now"                         "nuxt"                       
#> [317] "nuxtjs"                      "nwjs"                       
#> [319] "objective-c"                 "ocaml"                      
#> [321] "octave"                      "octobercms"                 
#> [323] "opa"                         "opencart"                   
#> [325] "opencv"                      "openfoam"                   
#> [327] "openframeworks"              "openframeworks+visualstudio"
#> [329] "oracleforms"                 "orcad"                      
#> [331] "osx"                         "otto"                       
#> [333] "oxideshop"                   "oxygenxmleditor"            
#> [335] "packer"                      "particle"                   
#> [337] "patch"                       "pawn"                       
#> [339] "perl"                        "perl6"                      
#> [341] "ph7cms"                      "phalcon"                    
#> [343] "phoenix"                     "phpcodesniffer"             
#> [345] "phpstorm"                    "phpstorm+all"               
#> [347] "phpstorm+iml"                "phpunit"                    
#> [349] "pico8"                       "pimcore"                    
#> [351] "pimcore4"                    "pimcore5"                   
#> [353] "pinegrow"                    "platformio"                 
#> [355] "playframework"               "plone"                      
#> [357] "polymer"                     "powershell"                 
#> [359] "premake-gmake"               "prepros"                    
#> [361] "prestashop"                  "processing"                 
#> [363] "progressabl"                 "psoccreator"                
#> [365] "puppet"                      "puppet-librarian"           
#> [367] "purebasic"                   "purescript"                 
#> [369] "putty"                       "pvs"                        
#> [371] "pycharm"                     "pycharm+all"                
#> [373] "pycharm+iml"                 "pydev"                      
#> [375] "python"                      "qml"                        
#> [377] "qooxdoo"                     "qt"                         
#> [379] "qtcreator"                   "r"                          
#> [381] "racket"                      "rails"                      
#> [383] "react"                       "reactnative"                
#> [385] "reasonml"                    "red"                        
#> [387] "redcar"                      "redis"                      
#> [389] "retool"                      "rhodesrhomobile"            
#> [391] "rider"                       "root"                       
#> [393] "ros"                         "ruby"                       
#> [395] "rubymine"                    "rubymine+all"               
#> [397] "rubymine+iml"                "rust"                       
#> [399] "salesforce"                  "salesforcedx"               
#> [401] "sas"                         "sass"                       
#> [403] "sbt"                         "scala"                      
#> [405] "scheme"                      "scons"                      
#> [407] "scrivener"                   "sdcc"                       
#> [409] "seamgen"                     "senchatouch"                
#> [411] "serverless"                  "shopware"                   
#> [413] "silverstripe"                "sketchup"                   
#> [415] "slickedit"                   "smalltalk"                  
#> [417] "snap"                        "snapcraft"                  
#> [419] "solidity"                    "soliditytruffle"            
#> [421] "sonar"                       "sonarqube"                  
#> [423] "sourcepawn"                  "spark"                      
#> [425] "splunk"                      "spreadsheet"                
#> [427] "ssh"                         "standardml"                 
#> [429] "stata"                       "stdlib"                     
#> [431] "stella"                      "stellar"                    
#> [433] "storybookjs"                 "stylus"                     
#> [435] "sublimetext"                 "sugarcrm"                   
#> [437] "svn"                         "swift"                      
#> [439] "swiftpackagemanager"         "swiftpm"                    
#> [441] "symfony"                     "symphonycms"                
#> [443] "synology"                    "synopsysvcs"                
#> [445] "tags"                        "tarmainstallmate"           
#> [447] "terraform"                   "terragrunt"                 
#> [449] "test"                        "testcomplete"               
#> [451] "testinfra"                   "tex"                        
#> [453] "text"                        "textmate"                   
#> [455] "textpattern"                 "theos-tweak"                
#> [457] "thinkphp"                    "tla+"                       
#> [459] "tortoisegit"                 "tower"                      
#> [461] "turbogears2"                 "twincat3"                   
#> [463] "tye"                         "typings"                    
#> [465] "typo3"                       "typo3-composer"             
#> [467] "umbraco"                     "unity"                      
#> [469] "unrealengine"                "vaadin"                     
#> [471] "vagrant"                     "valgrind"                   
#> [473] "vapor"                       "venv"                       
#> [475] "vertx"                       "video"                      
#> [477] "vim"                         "virtualenv"                 
#> [479] "virtuoso"                    "visualstudio"               
#> [481] "visualstudiocode"            "vivado"                     
#> [483] "vlab"                        "vs"                         
#> [485] "vscode"                      "vue"                        
#> [487] "vuejs"                       "vvvv"                       
#> [489] "waf"                         "wakanda"                    
#> [491] "web"                         "webmethods"                 
#> [493] "webstorm"                    "webstorm+all"               
#> [495] "webstorm+iml"                "werckercli"                 
#> [497] "windows"                     "wintersmith"                
#> [499] "wordpress"                   "wyam"                       
#> [501] "xamarinstudio"               "xcode"                      
#> [503] "xcodeinjection"              "xilinx"                     
#> [505] "xilinxise"                   "xilinxvivado"               
#> [507] "xill"                        "xojo"                       
#> [509] "xtext"                       "y86"                        
#> [511] "yarn"                        "yeoman"                     
#> [513] "yii"                         "yii2"                       
#> [515] "zendframework"               "zephir"                     
#> [517] "zig"                         "zsh"                        
#> [519] "zukencr8000"