Fetch available templates from gitignore.io

This return list of all templates supported by gitignore.io.

Usage

gi_available_templates()

Value

A character with all templates supported by gitignore.io.

Details

The returned list is returned as lower case characters.

Examples

gi_available_templates()
#>   [1] "1c"                          "1c-bitrix"                  
#>   [3] "a-frame"                     "al"                         
#>   [5] "aspnetcore"                  "ate"                        
#>   [7] "ats"                         "awr"                        
#>   [9] "actionscript"                "ada"                        
#>  [11] "adobe"                       "advancedinstaller"          
#>  [13] "adventuregamestudio"         "agda"                       
#>  [15] "alteraquartusii"             "altium"                     
#>  [17] "amplify"                     "android"                    
#>  [19] "androidstudio"               "angular"                    
#>  [21] "anjuta"                      "ansible"                    
#>  [23] "ansibletower"                "apachecordova"              
#>  [25] "apachehadoop"                "appbuilder"                 
#>  [27] "appcode"                     "appcode+all"                
#>  [29] "appcode+iml"                 "appengine"                  
#>  [31] "appceleratortitanium"        "aptanastudio"               
#>  [33] "arcanist"                    "archlinuxpackages"          
#>  [35] "archive"                     "archives"                   
#>  [37] "assembler"                   "astro"                      
#>  [39] "atmelstudio"                 "audio"                      
#>  [41] "autohotkey"                  "automationstudio"           
#>  [43] "autotools"                   "autotools+strict"           
#>  [45] "azurefunctions"              "azurite"                    
#>  [47] "backup"                      "ballerina"                  
#>  [49] "basic"                       "batch"                      
#>  [51] "bazaar"                      "bazel"                      
#>  [53] "bittorrent"                  "bitrise"                    
#>  [55] "bitrix"                      "blackbox"                   
#>  [57] "blender"                     "bloop"                      
#>  [59] "bookdown"                    "bower"                      
#>  [61] "bricxcc"                     "buck"                       
#>  [63] "c"                           "c++"                        
#>  [65] "cfwheels"                    "clion"                      
#>  [67] "clion+all"                   "clion+iml"                  
#>  [69] "cmake"                       "crbasic"                    
#>  [71] "cs-cart"                     "cuda"                       
#>  [73] "cvs"                         "cake"                       
#>  [75] "cakephp"                     "cakephp2"                   
#>  [77] "cakephp3"                    "calabash"                   
#>  [79] "carthage"                    "ceylon"                     
#>  [81] "chefcookbook"                "chocolatey"                 
#>  [83] "circuitpython"               "clean"                      
#>  [85] "clojure"                     "cloud9"                     
#>  [87] "cocoapods"                   "cocos2dx"                   
#>  [89] "cocoscreator"                "codeblocks"                 
#>  [91] "codecomposerstudio"          "codeigniter"                
#>  [93] "codekit"                     "codesniffer"                
#>  [95] "codeio"                      "coffeescript"               
#>  [97] "commonlisp"                  "compodoc"                   
#>  [99] "composer"                    "compressed"                 
#> [101] "compressedarchive"           "compression"                
#> [103] "conan"                       "concrete5"                  
#> [105] "coq"                         "cordova"                    
#> [107] "craftcms"                    "crashlytics"                
#> [109] "crossbar"                    "crystal"                    
#> [111] "csharp"                      "cypressio"                  
#> [113] "d"                           "dbeaver"                    
#> [115] "dm"                          "dart"                       
#> [117] "darteditor"                  "data"                       
#> [119] "datarecovery"                "database"                   
#> [121] "defold"                      "delphi"                     
#> [123] "deno"                        "dframe"                     
#> [125] "diff"                        "diskimage"                  
#> [127] "django"                      "docfx"                      
#> [129] "docpress"                    "docusaurus"                 
#> [131] "docz"                        "dotsettings"                
#> [133] "dotfilessh"                  "dotnetcore"                 
#> [135] "dreamweaver"                 "dropbox"                    
#> [137] "drupal"                      "drupal7"                    
#> [139] "drupal8"                     "episerver"                  
#> [141] "eagle"                       "eclipse"                    
#> [143] "eiffelstudio"                "elasticbeanstalk"           
#> [145] "elisp"                       "elixir"                     
#> [147] "elm"                         "emacs"                      
#> [149] "ember"                       "ensime"                     
#> [151] "erlang"                      "espresso"                   
#> [153] "executable"                  "exercism"                   
#> [155] "expressionengine"            "extjs"                      
#> [157] "fancy"                       "finale"                     
#> [159] "firebase"                    "fish"                       
#> [161] "flashbuilder"                "flask"                      
#> [163] "flatpak"                     "flex"                       
#> [165] "flexbuilder"                 "flutter"                    
#> [167] "font"                        "fontforge"                  
#> [169] "forcedotcom"                 "forgegradle"                
#> [171] "fortran"                     "freecad"                    
#> [173] "freepascal"                  "fuelphp"                    
#> [175] "fusetools"                   "ggts"                       
#> [177] "gis"                         "gpg"                        
#> [179] "gwt"                         "games"                      
#> [181] "gatsby"                      "gcov"                       
#> [183] "genero4gl"                   "geth"                       
#> [185] "git"                         "gitbook"                    
#> [187] "go"                          "goland"                     
#> [189] "goland+all"                  "goland+iml"                 
#> [191] "godot"                       "goodsync"                   
#> [193] "gradle"                      "grails"                     
#> [195] "groovy"                      "hol"                        
#> [197] "hsp"                         "haskell"                    
#> [199] "helm"                        "hexo"                       
#> [201] "homeassistant"               "homebrew"                   
#> [203] "hugo"                        "hyperledgercomposer"        
#> [205] "iar"                         "iarembeddedworkbench"       
#> [207] "iar_ewarm"                   "idapro"                     
#> [209] "igorpro"                     "idris"                      
#> [211] "images"                      "inforcms"                   
#> [213] "inforcrm"                    "intellij"                   
#> [215] "intellij+all"                "intellij+iml"               
#> [217] "ionic3"                      "jboss"                      
#> [219] "jboss-4-2-3-ga"              "jboss-6-x"                  
#> [221] "jboss4"                      "jboss6"                     
#> [223] "jdeveloper"                  "jenv"                       
#> [225] "jgiven"                      "jmeter"                     
#> [227] "jabref"                      "janet"                      
#> [229] "java"                        "jekyll"                     
#> [231] "jetbrains"                   "jetbrains+all"              
#> [233] "jetbrains+iml"               "jigsaw"                     
#> [235] "joe"                         "joomla"                     
#> [237] "jsonnet"                     "julia"                      
#> [239] "jupyternotebooks"            "justcode"                   
#> [241] "kdevelop4"                   "kdiff3"                     
#> [243] "kate"                        "keil"                       
#> [245] "kentico"                     "keystonejs"                 
#> [247] "kicad"                       "kirby2"                     
#> [249] "kirby3"                      "kobalt"                     
#> [251] "kohana"                      "komodoedit"                 
#> [253] "konyvisualizer"              "kotlin"                     
#> [255] "lamp"                        "lsspice"                    
#> [257] "ltspice"                     "latex"                      
#> [259] "labview"                     "labviewnxg"                 
#> [261] "laravel"                     "lazarus"                    
#> [263] "leiningen"                   "lemonstand"                 
#> [265] "less"                        "liberosoc"                  
#> [267] "libreoffice"                 "lighthouseci"               
#> [269] "lilypond"                    "linux"                      
#> [271] "lithium"                     "localstack"                 
#> [273] "logtalk"                     "lua"                        
#> [275] "lyx"                         "matlab"                     
#> [277] "mean"                        "modx"                       
#> [279] "mplabx"                      "magento"                    
#> [281] "magento1"                    "magento2"                   
#> [283] "magic-xpa"                   "maven"                      
#> [285] "mavensmate"                  "mdbook"                     
#> [287] "mercurial"                   "mercury"                    
#> [289] "meson"                       "metaprogrammingsystem"      
#> [291] "metals"                      "metalsmith"                 
#> [293] "meteor"                      "meteorjs"                   
#> [295] "microsoftoffice"             "mikroc"                     
#> [297] "mill"                        "moban"                      
#> [299] "modelsim"                    "momentics"                  
#> [301] "monodevelop"                 "ncrunch"                    
#> [303] "nwjs"                        "nanoc"                      
#> [305] "nativescript"                "nesc"                       
#> [307] "netbeans"                    "nette"                      
#> [309] "nextjs"                      "nikola"                     
#> [311] "nim"                         "ninja"                      
#> [313] "node"                        "nodechakratimetraveldebug"  
#> [315] "nohup"                       "notepadpp"                  
#> [317] "nuxtjs"                      "ocaml"                      
#> [319] "osx"                         "objective-c"                
#> [321] "obsidian"                    "octave"                     
#> [323] "octobercms"                  "opa"                        
#> [325] "opencv"                      "opencart"                   
#> [327] "openfoam"                    "openframeworks"             
#> [329] "openframeworks+visualstudio" "orcad"                      
#> [331] "oracleforms"                 "otto"                       
#> [333] "oxideshop"                   "pawn"                       
#> [335] "php-cs-fixer"                "phpcodesniffer"             
#> [337] "phpunit"                     "psoccreator"                
#> [339] "pvs"                         "packer"                     
#> [341] "pants"                       "particle"                   
#> [343] "patch"                       "perl"                       
#> [345] "perl6"                       "phalcon"                    
#> [347] "phoenix"                     "phpstorm"                   
#> [349] "phpstorm+all"                "phpstorm+iml"               
#> [351] "pimcore"                     "pimcore4"                   
#> [353] "pimcore5"                    "pinegrow"                   
#> [355] "platformio"                  "playframework"              
#> [357] "plone"                       "polymer"                    
#> [359] "powershell"                  "prepros"                    
#> [361] "prestashop"                  "processing"                 
#> [363] "progressabl"                 "putty"                      
#> [365] "pulumi"                      "pulumi+stacks"              
#> [367] "puppet"                      "purebasic"                  
#> [369] "purescript"                  "pycharm"                    
#> [371] "pycharm+all"                 "pycharm+iml"                
#> [373] "python"                      "pythonvanilla"              
#> [375] "qml"                         "qooxdoo"                    
#> [377] "qt"                          "qtcreator"                  
#> [379] "r"                           "root"                       
#> [381] "ros"                         "ros2"                       
#> [383] "racket"                      "rails"                      
#> [385] "reactnative"                 "reasonml"                   
#> [387] "red"                         "redcar"                     
#> [389] "redis"                       "remix"                      
#> [391] "remix+arc"                   "remix+cloudflarepages"      
#> [393] "remix+cloudflareworkers"     "remix+netlify"              
#> [395] "remix+vercel"                "renpy"                      
#> [397] "rhodesrhomobile"             "rider"                      
#> [399] "ruby"                        "rubymine"                   
#> [401] "rubymine+all"                "rubymine+iml"               
#> [403] "rust"                        "sam"                        
#> [405] "sam+config"                  "sas"                        
#> [407] "sbt"                         "scons"                      
#> [409] "ssh"                         "svn"                        
#> [411] "salesforce"                  "salesforcedx"               
#> [413] "sass"                        "scala"                      
#> [415] "scheme"                      "scrivener"                  
#> [417] "sdcc"                        "seamgen"                    
#> [419] "senchatouch"                 "serverless"                 
#> [421] "shopware"                    "silverstripe"               
#> [423] "sketchup"                    "slickedit"                  
#> [425] "smalltalk"                   "snap"                       
#> [427] "snapcraft"                   "snyk"                       
#> [429] "solidity"                    "soliditytruffle"            
#> [431] "sonar"                       "sonarqube"                  
#> [433] "sourcepawn"                  "spark"                      
#> [435] "specflow"                    "splunk"                     
#> [437] "spreadsheet"                 "standardml"                 
#> [439] "stata"                       "stdlib"                     
#> [441] "stella"                      "stellar"                    
#> [443] "storybookjs"                 "strapi"                     
#> [445] "stylus"                      "sublimetext"                
#> [447] "sugarcrm"                    "svelte"                     
#> [449] "swift"                       "swiftpm"                    
#> [451] "swiftpackagemanager"         "symfony"                    
#> [453] "symphonycms"                 "synology"                   
#> [455] "synopsysvcs"                 "theos-tweak"                
#> [457] "tla+"                        "typo3-composer"             
#> [459] "tags"                        "tarmainstallmate"           
#> [461] "tex"                         "terraform"                  
#> [463] "terragrunt"                  "test"                       
#> [465] "testcomplete"                "testinfra"                  
#> [467] "text"                        "textmate"                   
#> [469] "textpattern"                 "thinkphp"                   
#> [471] "toit"                        "tortoisegit"                
#> [473] "tower"                       "turbo"                      
#> [475] "turbogears2"                 "twincat3"                   
#> [477] "tye"                         "typings"                    
#> [479] "typo3"                       "umbraco"                    
#> [481] "unity"                       "unrealengine"               
#> [483] "vlab"                        "vvvv"                       
#> [485] "vaadin"                      "vagrant"                    
#> [487] "valgrind"                    "vapor"                      
#> [489] "vercel"                      "vertx"                      
#> [491] "video"                       "vim"                        
#> [493] "virtualenv"                  "virtuoso"                   
#> [495] "visualbasic"                 "visualstudio"               
#> [497] "visualstudiocode"            "vivado"                     
#> [499] "vue"                         "vuejs"                      
#> [501] "waf"                         "wakanda"                    
#> [503] "web"                         "webmethods"                 
#> [505] "webstorm"                    "webstorm+all"               
#> [507] "webstorm+iml"                "werckercli"                 
#> [509] "windows"                     "wintersmith"                
#> [511] "wordpress"                   "wyam"                       
#> [513] "xtext"                       "xamarinstudio"              
#> [515] "xcode"                       "xcodeinjection"             
#> [517] "xilinx"                      "xilinxise"                  
#> [519] "xilinxvivado"                "xill"                       
#> [521] "xojo"                        "y86"                        
#> [523] "yeoman"                      "yii"                        
#> [525] "yii2"                        "zendframework"              
#> [527] "zephir"                      "zsh"                        
#> [529] "zukencr8000"                 "asdf"                       
#> [531] "basercms"                    "bluej"                      
#> [533] "certificates"                "dbt"                        
#> [535] "direnv"                      "dotenv"                     
#> [537] "e2studio"                    "easybook"                   
#> [539] "fastlane"                    "floobits"                   
#> [541] "fsharp"                      "greenfoot"                  
#> [543] "grunt"                       "infer"                      
#> [545] "jspm"                        "kaldi"                      
#> [547] "librarian-chef"              "macos"                      
#> [549] "mule"                        "nova"                       
#> [551] "now"                         "oxygenxmleditor"            
#> [553] "ph7cms"                      "pico8"                      
#> [555] "premake-gmake"               "puppet-librarian"           
#> [557] "pydev"                       "react"                      
#> [559] "replit"                      "retool"                     
#> [561] "robotframework"              "rust-analyzer"              
#> [563] "vrealizeorchestrator"        "vcpkg"                      
#> [565] "venv"                        "vs"                         
#> [567] "yalc"                        "yarn"                       
#> [569] "zig"