Fetch available templates from gitignore.io

This return list of all templates supported by gitignore.io.

Usage

gi_available_templates()

Value

A character with all templates supported by gitignore.io.

Details

The returned list is returned as lower case characters.

Examples

gi_available_templates()
#>   [1] "1c"                          "1c-bitrix"                  
#>   [3] "a-frame"                     "al"                         
#>   [5] "aspnetcore"                  "ate"                        
#>   [7] "ats"                         "awr"                        
#>   [9] "actionscript"                "ada"                        
#>  [11] "adobe"                       "advancedinstaller"          
#>  [13] "adventuregamestudio"         "agda"                       
#>  [15] "alteraquartusii"             "altium"                     
#>  [17] "amplify"                     "android"                    
#>  [19] "androidstudio"               "angular"                    
#>  [21] "anjuta"                      "ansible"                    
#>  [23] "ansibletower"                "apachecordova"              
#>  [25] "apachehadoop"                "appbuilder"                 
#>  [27] "appcode"                     "appcode+all"                
#>  [29] "appcode+iml"                 "appengine"                  
#>  [31] "appceleratortitanium"        "aptanastudio"               
#>  [33] "arcanist"                    "archlinuxpackages"          
#>  [35] "archive"                     "archives"                   
#>  [37] "assembler"                   "astro"                      
#>  [39] "atmelstudio"                 "audio"                      
#>  [41] "autohotkey"                  "automationstudio"           
#>  [43] "autotools"                   "autotools+strict"           
#>  [45] "azurefunctions"              "azurite"                    
#>  [47] "backup"                      "ballerina"                  
#>  [49] "basic"                       "batch"                      
#>  [51] "bazaar"                      "bazel"                      
#>  [53] "bittorrent"                  "bitrise"                    
#>  [55] "bitrix"                      "blackbox"                   
#>  [57] "blender"                     "bloop"                      
#>  [59] "bookdown"                    "bower"                      
#>  [61] "bricxcc"                     "buck"                       
#>  [63] "c"                           "c++"                        
#>  [65] "cfwheels"                    "clion"                      
#>  [67] "clion+all"                   "clion+iml"                  
#>  [69] "cmake"                       "crbasic"                    
#>  [71] "cs-cart"                     "cuda"                       
#>  [73] "cvs"                         "cake"                       
#>  [75] "cakephp"                     "cakephp2"                   
#>  [77] "cakephp3"                    "calabash"                   
#>  [79] "carthage"                    "ceylon"                     
#>  [81] "chefcookbook"                "chocolatey"                 
#>  [83] "circuitpython"               "clean"                      
#>  [85] "clojure"                     "cloud9"                     
#>  [87] "cocoapods"                   "cocos2dx"                   
#>  [89] "cocoscreator"                "codeblocks"                 
#>  [91] "codecomposerstudio"          "codeigniter"                
#>  [93] "codekit"                     "codesniffer"                
#>  [95] "codeio"                      "coffeescript"               
#>  [97] "commonlisp"                  "compodoc"                   
#>  [99] "composer"                    "compressed"                 
#> [101] "compressedarchive"           "compression"                
#> [103] "conan"                       "concrete5"                  
#> [105] "coq"                         "cordova"                    
#> [107] "craftcms"                    "crashlytics"                
#> [109] "crossbar"                    "crystal"                    
#> [111] "csharp"                      "cypressio"                  
#> [113] "d"                           "dbeaver"                    
#> [115] "dm"                          "dart"                       
#> [117] "darteditor"                  "data"                       
#> [119] "datarecovery"                "database"                   
#> [121] "defold"                      "delphi"                     
#> [123] "deno"                        "dframe"                     
#> [125] "diff"                        "diskimage"                  
#> [127] "django"                      "docfx"                      
#> [129] "docpress"                    "docusaurus"                 
#> [131] "docz"                        "dotsettings"                
#> [133] "dotfilessh"                  "dotnetcore"                 
#> [135] "doxygen"                     "dreamweaver"                
#> [137] "dropbox"                     "drupal"                     
#> [139] "drupal7"                     "drupal8"                    
#> [141] "episerver"                   "eagle"                      
#> [143] "eclipse"                     "eiffelstudio"               
#> [145] "elasticbeanstalk"            "elisp"                      
#> [147] "elixir"                      "elm"                        
#> [149] "emacs"                       "ember"                      
#> [151] "ensime"                      "erlang"                     
#> [153] "espresso"                    "executable"                 
#> [155] "exercism"                    "expressionengine"           
#> [157] "extjs"                       "fancy"                      
#> [159] "finale"                      "firebase"                   
#> [161] "fish"                        "flashbuilder"               
#> [163] "flask"                       "flatpak"                    
#> [165] "flex"                        "flexbuilder"                
#> [167] "flutter"                     "font"                       
#> [169] "fontforge"                   "forcedotcom"                
#> [171] "forgegradle"                 "fortran"                    
#> [173] "freecad"                     "freepascal"                 
#> [175] "fuelphp"                     "fusetools"                  
#> [177] "ggts"                        "gis"                        
#> [179] "gpg"                         "gwt"                        
#> [181] "games"                       "gatsby"                     
#> [183] "gcov"                        "genero4gl"                  
#> [185] "geth"                        "git"                        
#> [187] "gitbook"                     "go"                         
#> [189] "goland"                      "goland+all"                 
#> [191] "goland+iml"                  "godot"                      
#> [193] "goodsync"                    "gradle"                     
#> [195] "grails"                      "groovy"                     
#> [197] "hol"                         "hsp"                        
#> [199] "haskell"                     "helm"                       
#> [201] "hexo"                        "homeassistant"              
#> [203] "homebrew"                    "hugo"                       
#> [205] "hyperledgercomposer"         "iar"                        
#> [207] "iarembeddedworkbench"        "iar_ewarm"                  
#> [209] "idapro"                      "igorpro"                    
#> [211] "idris"                       "images"                     
#> [213] "inforcms"                    "inforcrm"                   
#> [215] "intellij"                    "intellij+all"               
#> [217] "intellij+iml"                "ionic3"                     
#> [219] "jboss"                       "jboss-4-2-3-ga"             
#> [221] "jboss-6-x"                   "jboss4"                     
#> [223] "jboss6"                      "jdeveloper"                 
#> [225] "jenv"                        "jgiven"                     
#> [227] "jmeter"                      "jabref"                     
#> [229] "janet"                       "java"                       
#> [231] "jekyll"                      "jetbrains"                  
#> [233] "jetbrains+all"               "jetbrains+iml"              
#> [235] "jigsaw"                      "joe"                        
#> [237] "joomla"                      "jsonnet"                    
#> [239] "julia"                       "jupyternotebooks"           
#> [241] "justcode"                    "kdevelop4"                  
#> [243] "kdiff3"                      "kate"                       
#> [245] "keil"                        "kentico"                    
#> [247] "keystonejs"                  "kicad"                      
#> [249] "kirby2"                      "kirby3"                     
#> [251] "kobalt"                      "kohana"                     
#> [253] "komodoedit"                  "konyvisualizer"             
#> [255] "kotlin"                      "lamp"                       
#> [257] "lsspice"                     "ltspice"                    
#> [259] "latex"                       "labview"                    
#> [261] "labviewnxg"                  "laravel"                    
#> [263] "lazarus"                     "leiningen"                  
#> [265] "lemonstand"                  "less"                       
#> [267] "liberosoc"                   "libreoffice"                
#> [269] "lighthouseci"                "lilypond"                   
#> [271] "linux"                       "lithium"                    
#> [273] "localstack"                  "logtalk"                    
#> [275] "lua"                         "lyx"                        
#> [277] "matlab"                      "mean"                       
#> [279] "modx"                        "mplabx"                     
#> [281] "magento"                     "magento1"                   
#> [283] "magento2"                    "magic-xpa"                  
#> [285] "maven"                       "mavensmate"                 
#> [287] "mdbook"                      "mercurial"                  
#> [289] "mercury"                     "meson"                      
#> [291] "metaprogrammingsystem"       "metals"                     
#> [293] "metalsmith"                  "meteor"                     
#> [295] "meteorjs"                    "microsoftoffice"            
#> [297] "mikroc"                      "mill"                       
#> [299] "moban"                       "modelsim"                   
#> [301] "momentics"                   "monodevelop"                
#> [303] "ncrunch"                     "nwjs"                       
#> [305] "nanoc"                       "nativescript"               
#> [307] "nesc"                        "netbeans"                   
#> [309] "nette"                       "nextjs"                     
#> [311] "nikola"                      "nim"                        
#> [313] "ninja"                       "node"                       
#> [315] "nodechakratimetraveldebug"   "nohup"                      
#> [317] "notepadpp"                   "nuxtjs"                     
#> [319] "ocaml"                       "osx"                        
#> [321] "objective-c"                 "obsidian"                   
#> [323] "octave"                      "octobercms"                 
#> [325] "opa"                         "opencv"                     
#> [327] "opencart"                    "openfoam"                   
#> [329] "openframeworks"              "openframeworks+visualstudio"
#> [331] "orcad"                       "oracleforms"                
#> [333] "otto"                        "oxideshop"                  
#> [335] "pawn"                        "php-cs-fixer"               
#> [337] "phpcodesniffer"              "phpunit"                    
#> [339] "psoccreator"                 "pvs"                        
#> [341] "packer"                      "pants"                      
#> [343] "particle"                    "patch"                      
#> [345] "perl"                        "perl6"                      
#> [347] "phalcon"                     "phoenix"                    
#> [349] "phpstorm"                    "phpstorm+all"               
#> [351] "phpstorm+iml"                "pimcore"                    
#> [353] "pimcore4"                    "pimcore5"                   
#> [355] "pinegrow"                    "platformio"                 
#> [357] "playframework"               "plone"                      
#> [359] "polymer"                     "powershell"                 
#> [361] "prepros"                     "prestashop"                 
#> [363] "processing"                  "progressabl"                
#> [365] "putty"                       "pulumi"                     
#> [367] "pulumi+stacks"               "puppet"                     
#> [369] "purebasic"                   "purescript"                 
#> [371] "pycharm"                     "pycharm+all"                
#> [373] "pycharm+iml"                 "python"                     
#> [375] "pythonvanilla"               "qml"                        
#> [377] "qooxdoo"                     "qt"                         
#> [379] "qtcreator"                   "r"                          
#> [381] "root"                        "ros"                        
#> [383] "ros2"                        "racket"                     
#> [385] "rails"                       "reactnative"                
#> [387] "reasonml"                    "red"                        
#> [389] "redcar"                      "redis"                      
#> [391] "remix"                       "remix+arc"                  
#> [393] "remix+cloudflarepages"       "remix+cloudflareworkers"    
#> [395] "remix+netlify"               "remix+vercel"               
#> [397] "renpy"                       "rhodesrhomobile"            
#> [399] "rider"                       "ruby"                       
#> [401] "rubymine"                    "rubymine+all"               
#> [403] "rubymine+iml"                "rust"                       
#> [405] "sam"                         "sam+config"                 
#> [407] "sas"                         "sbt"                        
#> [409] "scons"                       "ssh"                        
#> [411] "svn"                         "salesforce"                 
#> [413] "salesforcedx"                "sass"                       
#> [415] "scala"                       "scheme"                     
#> [417] "scrivener"                   "sdcc"                       
#> [419] "seamgen"                     "senchatouch"                
#> [421] "serverless"                  "shopware"                   
#> [423] "silverstripe"                "sketchup"                   
#> [425] "slickedit"                   "smalltalk"                  
#> [427] "snap"                        "snapcraft"                  
#> [429] "snyk"                        "solidity"                   
#> [431] "soliditytruffle"             "sonar"                      
#> [433] "sonarqube"                   "sourcepawn"                 
#> [435] "spark"                       "specflow"                   
#> [437] "splunk"                      "spreadsheet"                
#> [439] "standardml"                  "stata"                      
#> [441] "stdlib"                      "stella"                     
#> [443] "stellar"                     "storybookjs"                
#> [445] "strapi"                      "stylus"                     
#> [447] "sublimetext"                 "sugarcrm"                   
#> [449] "svelte"                      "swift"                      
#> [451] "swiftpm"                     "swiftpackagemanager"        
#> [453] "symfony"                     "symphonycms"                
#> [455] "synology"                    "synopsysvcs"                
#> [457] "theos-tweak"                 "tla+"                       
#> [459] "typo3-composer"              "tags"                       
#> [461] "tarmainstallmate"            "tex"                        
#> [463] "terraform"                   "terragrunt"                 
#> [465] "test"                        "testcomplete"               
#> [467] "testinfra"                   "text"                       
#> [469] "textmate"                    "textpattern"                
#> [471] "thinkphp"                    "toit"                       
#> [473] "tortoisegit"                 "tower"                      
#> [475] "turbo"                       "turbogears2"                
#> [477] "twincat3"                    "tye"                        
#> [479] "typings"                     "typo3"                      
#> [481] "umbraco"                     "unity"                      
#> [483] "unrealengine"                "vlab"                       
#> [485] "vvvv"                        "vaadin"                     
#> [487] "vagrant"                     "valgrind"                   
#> [489] "vapor"                       "vercel"                     
#> [491] "vertx"                       "video"                      
#> [493] "vim"                         "virtualenv"                 
#> [495] "virtuoso"                    "visualbasic"                
#> [497] "visualstudio"                "visualstudiocode"           
#> [499] "vivado"                      "vue"                        
#> [501] "vuejs"                       "waf"                        
#> [503] "wakanda"                     "web"                        
#> [505] "webmethods"                  "webstorm"                   
#> [507] "webstorm+all"                "webstorm+iml"               
#> [509] "werckercli"                  "windows"                    
#> [511] "wintersmith"                 "wordpress"                  
#> [513] "wyam"                        "xtext"                      
#> [515] "xamarinstudio"               "xcode"                      
#> [517] "xcodeinjection"              "xilinx"                     
#> [519] "xilinxise"                   "xilinxvivado"               
#> [521] "xill"                        "xojo"                       
#> [523] "y86"                         "yeoman"                     
#> [525] "yii"                         "yii2"                       
#> [527] "zendframework"               "zephir"                     
#> [529] "zsh"                         "zukencr8000"                
#> [531] "asdf"                        "basercms"                   
#> [533] "bluej"                       "certificates"               
#> [535] "dbt"                         "direnv"                     
#> [537] "dotenv"                      "e2studio"                   
#> [539] "easybook"                    "fastlane"                   
#> [541] "floobits"                    "fsharp"                     
#> [543] "greenfoot"                   "grunt"                      
#> [545] "infer"                       "jspm"                       
#> [547] "kaldi"                       "librarian-chef"             
#> [549] "macos"                       "mule"                       
#> [551] "nova"                        "now"                        
#> [553] "oxygenxmleditor"             "ph7cms"                     
#> [555] "pico8"                       "premake-gmake"              
#> [557] "puppet-librarian"            "pydev"                      
#> [559] "react"                       "replit"                     
#> [561] "retool"                      "robotframework"             
#> [563] "rust-analyzer"               "vrealizeorchestrator"       
#> [565] "vcpkg"                       "venv"                       
#> [567] "vs"                          "xmake"                      
#> [569] "yalc"                        "yarn"                       
#> [571] "zig"