Fetch available templates from gitignore.io

This return list of all templates supported by gitignore.io.

gi_available_templates()

Value

A character with all templates supported by gitignore.io.

Details

The returned list is returned as lower case characters.

Examples

gi_available_templates()

#>   [1] "1c"                          "1c-bitrix"                  
#>   [3] "a-frame"                     "actionscript"               
#>   [5] "ada"                         "adobe"                      
#>   [7] "advancedinstaller"           "adventuregamestudio"        
#>   [9] "agda"                        "al"                         
#>  [11] "alteraquartusii"             "altium"                     
#>  [13] "android"                     "androidstudio"              
#>  [15] "angular"                     "anjuta"                     
#>  [17] "ansible"                     "apachecordova"              
#>  [19] "apachehadoop"                "appbuilder"                 
#>  [21] "appceleratortitanium"        "appcode"                    
#>  [23] "appcode+all"                 "appcode+iml"                
#>  [25] "appengine"                   "aptanastudio"               
#>  [27] "arcanist"                    "archive"                    
#>  [29] "archives"                    "archlinuxpackages"          
#>  [31] "aspnetcore"                  "assembler"                  
#>  [33] "ate"                         "atmelstudio"                
#>  [35] "ats"                         "audio"                      
#>  [37] "automationstudio"            "autotools"                  
#>  [39] "autotools+strict"            "awr"                        
#>  [41] "backup"                      "ballerina"                  
#>  [43] "basercms"                    "basic"                      
#>  [45] "batch"                       "bazaar"                     
#>  [47] "bazel"                       "bitrise"                    
#>  [49] "bitrix"                      "bittorrent"                 
#>  [51] "blackbox"                    "bloop"                      
#>  [53] "bluej"                       "bookdown"                   
#>  [55] "bower"                       "bricxcc"                    
#>  [57] "buck"                        "c"                          
#>  [59] "c++"                         "cake"                       
#>  [61] "cakephp"                     "cakephp2"                   
#>  [63] "cakephp3"                    "calabash"                   
#>  [65] "carthage"                    "ceylon"                     
#>  [67] "cfwheels"                    "chefcookbook"               
#>  [69] "chocolatey"                  "clean"                      
#>  [71] "clion"                       "clion+all"                  
#>  [73] "clion+iml"                   "clojure"                    
#>  [75] "cloud9"                      "cmake"                      
#>  [77] "cocoapods"                   "cocos2dx"                   
#>  [79] "cocoscreator"                "code"                       
#>  [81] "code-java"                   "codeblocks"                 
#>  [83] "codecomposerstudio"          "codeigniter"                
#>  [85] "codeio"                      "codekit"                    
#>  [87] "codesniffer"                 "coffeescript"               
#>  [89] "commonlisp"                  "composer"                   
#>  [91] "compressed"                  "compressedarchive"          
#>  [93] "compression"                 "concrete5"                  
#>  [95] "coq"                         "cordova"                    
#>  [97] "craftcms"                    "crashlytics"                
#>  [99] "crbasic"                     "crossbar"                   
#> [101] "crystal"                     "csharp"                     
#> [103] "cuda"                        "cvs"                        
#> [105] "d"                           "dart"                       
#> [107] "darteditor"                  "data"                       
#> [109] "database"                    "datarecovery"               
#> [111] "dbeaver"                     "defold"                     
#> [113] "delphi"                      "dframe"                     
#> [115] "diff"                        "diskimage"                  
#> [117] "django"                      "dm"                         
#> [119] "docfx"                       "docpress"                   
#> [121] "dotfilessh"                  "dotnetcore"                 
#> [123] "dotsettings"                 "dreamweaver"                
#> [125] "dropbox"                     "drupal"                     
#> [127] "drupal7"                     "drupal8"                    
#> [129] "eagle"                       "easybook"                   
#> [131] "eclipse"                     "eiffelstudio"               
#> [133] "elasticbeanstalk"            "elisp"                      
#> [135] "elixir"                      "elm"                        
#> [137] "emacs"                       "ember"                      
#> [139] "ensime"                      "episerver"                  
#> [141] "erlang"                      "espresso"                   
#> [143] "executable"                  "exercism"                   
#> [145] "expressionengine"            "extjs"                      
#> [147] "fancy"                       "fastlane"                   
#> [149] "finale"                      "firebase"                   
#> [151] "flashbuilder"                "flask"                      
#> [153] "flex"                        "flexbuilder"                
#> [155] "floobits"                    "flutter"                    
#> [157] "font"                        "fontforge"                  
#> [159] "forcedotcom"                 "forgegradle"                
#> [161] "fortran"                     "freepascal"                 
#> [163] "fsharp"                      "fuelphp"                    
#> [165] "fusetools"                   "games"                      
#> [167] "gcov"                        "genero4gl"                  
#> [169] "geth"                        "ggts"                       
#> [171] "gis"                         "git"                        
#> [173] "gitbook"                     "go"                         
#> [175] "godot"                       "goodsync"                   
#> [177] "gpg"                         "gradle"                     
#> [179] "grails"                      "greenfoot"                  
#> [181] "grunt"                       "gwt"                        
#> [183] "haskell"                     "helm"                       
#> [185] "hexo"                        "hol"                        
#> [187] "homeassistant"               "hsp"                        
#> [189] "hugo"                        "hyperledgercomposer"        
#> [191] "iar"                         "iar_ewarm"                  
#> [193] "iarembeddedworkbench"        "idapro"                     
#> [195] "idris"                       "igorpro"                    
#> [197] "images"                      "infer"                      
#> [199] "inforcms"                    "inforcrm"                   
#> [201] "intellij"                    "intellij+all"               
#> [203] "intellij+iml"                "ionic3"                     
#> [205] "jabref"                      "java"                       
#> [207] "java-web"                    "jboss"                      
#> [209] "jboss-4-2-3-ga"              "jboss-6-x"                  
#> [211] "jboss4"                      "jboss6"                     
#> [213] "jdeveloper"                  "jekyll"                     
#> [215] "jenv"                        "jetbrains"                  
#> [217] "jetbrains+all"               "jetbrains+iml"              
#> [219] "jgiven"                      "jigsaw"                     
#> [221] "jmeter"                      "joe"                        
#> [223] "joomla"                      "jspm"                       
#> [225] "julia"                       "jupyternotebooks"           
#> [227] "justcode"                    "kate"                       
#> [229] "kdevelop4"                   "kdiff3"                     
#> [231] "keil"                        "kentico"                    
#> [233] "kicad"                       "kirby2"                     
#> [235] "kobalt"                      "kohana"                     
#> [237] "komodoedit"                  "konyvisualizer"             
#> [239] "kotlin"                      "labview"                    
#> [241] "labviewnxg"                  "lamp"                       
#> [243] "laravel"                     "latex"                      
#> [245] "lazarus"                     "leiningen"                  
#> [247] "lemonstand"                  "less"                       
#> [249] "liberosoc"                   "librarian-chef"             
#> [251] "libreoffice"                 "lilypond"                   
#> [253] "linux"                       "lithium"                    
#> [255] "logtalk"                     "lsspice"                    
#> [257] "ltspice"                     "lua"                        
#> [259] "lyx"                         "m2e"                        
#> [261] "macos"                       "magento"                    
#> [263] "magento1"                    "magento2"                   
#> [265] "magic-xpa"                   "matlab"                     
#> [267] "maven"                       "mavensmate"                 
#> [269] "mdbook"                      "mean"                       
#> [271] "mercurial"                   "mercury"                    
#> [273] "metals"                      "metaprogrammingsystem"      
#> [275] "meteor"                      "meteorjs"                   
#> [277] "microsoftoffice"             "mikroc"                     
#> [279] "moban"                       "modelsim"                   
#> [281] "modx"                        "momentics"                  
#> [283] "monodevelop"                 "mplabx"                     
#> [285] "mule"                        "nanoc"                      
#> [287] "nativescript"                "ncrunch"                    
#> [289] "nesc"                        "netbeans"                   
#> [291] "nette"                       "nikola"                     
#> [293] "nim"                         "ninja"                      
#> [295] "node"                        "nodechakratimetraveldebug"  
#> [297] "notepadpp"                   "nuxt"                       
#> [299] "nuxtjs"                      "nwjs"                       
#> [301] "objective-c"                 "ocaml"                      
#> [303] "octave"                      "octobercms"                 
#> [305] "opa"                         "opencart"                   
#> [307] "opencv"                      "openfoam"                   
#> [309] "openframeworks"              "openframeworks+visualstudio"
#> [311] "oracleforms"                 "orcad"                      
#> [313] "osx"                         "otto"                       
#> [315] "oxideshop"                   "packer"                     
#> [317] "particle"                    "patch"                      
#> [319] "pawn"                        "perl"                       
#> [321] "perl6"                       "ph7cms"                     
#> [323] "phalcon"                     "phoenix"                    
#> [325] "phpcodesniffer"              "phpstorm"                   
#> [327] "phpstorm+all"                "phpstorm+iml"               
#> [329] "pico8"                       "pimcore"                    
#> [331] "pimcore4"                    "pimcore5"                   
#> [333] "pinegrow"                    "platformio"                 
#> [335] "playframework"               "plone"                      
#> [337] "polymer"                     "powershell"                 
#> [339] "premake-gmake"               "prepros"                    
#> [341] "prestashop"                  "processing"                 
#> [343] "progressabl"                 "psoccreator"                
#> [345] "puppet"                      "puppet-librarian"           
#> [347] "purebasic"                   "purescript"                 
#> [349] "putty"                       "pvs"                        
#> [351] "pycharm"                     "pycharm+all"                
#> [353] "pycharm+iml"                 "pydev"                      
#> [355] "python"                      "qml"                        
#> [357] "qooxdoo"                     "qt"                         
#> [359] "qtcreator"                   "r"                          
#> [361] "racket"                      "rails"                      
#> [363] "reactnative"                 "reasonml"                   
#> [365] "red"                         "redcar"                     
#> [367] "redis"                       "rhodesrhomobile"            
#> [369] "rider"                       "root"                       
#> [371] "ros"                         "ruby"                       
#> [373] "rubymine"                    "rubymine+all"               
#> [375] "rubymine+iml"                "rust"                       
#> [377] "salesforce"                  "salesforcedx"               
#> [379] "sas"                         "sass"                       
#> [381] "sbt"                         "scala"                      
#> [383] "scheme"                      "scons"                      
#> [385] "scrivener"                   "sdcc"                       
#> [387] "seamgen"                     "senchatouch"                
#> [389] "serverless"                  "shopware"                   
#> [391] "silverstripe"                "sketchup"                   
#> [393] "slickedit"                   "smalltalk"                  
#> [395] "snap"                        "snapcraft"                  
#> [397] "solidity"                    "soliditytruffle"            
#> [399] "sonar"                       "sonarqube"                  
#> [401] "sourcepawn"                  "spark"                      
#> [403] "splunk"                      "spreadsheet"                
#> [405] "standardml"                  "stata"                      
#> [407] "stdlib"                      "stella"                     
#> [409] "stellar"                     "stylus"                     
#> [411] "sublimetext"                 "sugarcrm"                   
#> [413] "svn"                         "swift"                      
#> [415] "swiftpackagemanager"         "swiftpm"                    
#> [417] "symfony"                     "symphonycms"                
#> [419] "synology"                    "synopsysvcs"                
#> [421] "tags"                        "tarmainstallmate"           
#> [423] "terraform"                   "test"                       
#> [425] "testcomplete"                "testinfra"                  
#> [427] "tex"                         "text"                       
#> [429] "textmate"                    "textpattern"                
#> [431] "theos-tweak"                 "thinkphp"                   
#> [433] "tortoisegit"                 "tower"                      
#> [435] "turbogears2"                 "twincat"                    
#> [437] "typings"                     "typo3"                      
#> [439] "typo3-composer"              "umbraco"                    
#> [441] "unity"                       "unrealengine"               
#> [443] "vaadin"                      "vagrant"                    
#> [445] "valgrind"                    "vapor"                      
#> [447] "venv"                        "vertx"                      
#> [449] "video"                       "vim"                        
#> [451] "virtualenv"                  "virtuoso"                   
#> [453] "visualstudio"                "visualstudiocode"           
#> [455] "vivado"                      "vlab"                       
#> [457] "vue"                         "vuejs"                      
#> [459] "vvvv"                        "waf"                        
#> [461] "wakanda"                     "web"                        
#> [463] "webmethods"                  "webstorm"                   
#> [465] "webstorm+all"                "webstorm+iml"               
#> [467] "werckercli"                  "windows"                    
#> [469] "wintersmith"                 "wordpress"                  
#> [471] "wyam"                        "xamarinstudio"              
#> [473] "xcode"                       "xcodeinjection"             
#> [475] "xilinx"                      "xilinxise"                  
#> [477] "xilinxvivado"                "xill"                       
#> [479] "xojo"                        "xtext"                      
#> [481] "y86"                         "yeoman"                     
#> [483] "yii"                         "yii2"                       
#> [485] "zendframework"               "zephir"                     
#> [487] "zig"                         "zukencr8000"