Fetch available templates from gitignore.io

This return list of all templates supported by gitignore.io.

gi_available_templates()

Value

A character with all templates supported by gitignore.io.

Details

The returned list is returned as lower case characters.

Examples

gi_available_templates()

#>   [1] "1c"                          "1c-bitrix"                  
#>   [3] "a-frame"                     "actionscript"               
#>   [5] "ada"                         "adobe"                      
#>   [7] "advancedinstaller"           "adventuregamestudio"        
#>   [9] "agda"                        "al"                         
#>  [11] "alteraquartusii"             "altium"                     
#>  [13] "android"                     "androidstudio"              
#>  [15] "angular"                     "anjuta"                     
#>  [17] "ansible"                     "apachecordova"              
#>  [19] "apachehadoop"                "appbuilder"                 
#>  [21] "appceleratortitanium"        "appcode"                    
#>  [23] "appcode+all"                 "appcode+iml"                
#>  [25] "appengine"                   "aptanastudio"               
#>  [27] "arcanist"                    "archive"                    
#>  [29] "archives"                    "archlinuxpackages"          
#>  [31] "aspnetcore"                  "assembler"                  
#>  [33] "ate"                         "atmelstudio"                
#>  [35] "ats"                         "audio"                      
#>  [37] "automationstudio"            "autotools"                  
#>  [39] "autotools+strict"            "awr"                        
#>  [41] "backup"                      "ballerina"                  
#>  [43] "basercms"                    "basic"                      
#>  [45] "batch"                       "bazaar"                     
#>  [47] "bazel"                       "bitrise"                    
#>  [49] "bitrix"                      "bittorrent"                 
#>  [51] "blackbox"                    "bloop"                      
#>  [53] "bluej"                       "bookdown"                   
#>  [55] "bower"                       "bricxcc"                    
#>  [57] "buck"                        "c"                          
#>  [59] "c++"                         "cake"                       
#>  [61] "cakephp"                     "cakephp2"                   
#>  [63] "cakephp3"                    "calabash"                   
#>  [65] "carthage"                    "certificates"               
#>  [67] "ceylon"                      "cfwheels"                   
#>  [69] "chefcookbook"                "chocolatey"                 
#>  [71] "clean"                       "clion"                      
#>  [73] "clion+all"                   "clion+iml"                  
#>  [75] "clojure"                     "cloud9"                     
#>  [77] "cmake"                       "cocoapods"                  
#>  [79] "cocos2dx"                    "cocoscreator"               
#>  [81] "code"                        "code-java"                  
#>  [83] "codeblocks"                  "codecomposerstudio"         
#>  [85] "codeigniter"                 "codeio"                     
#>  [87] "codekit"                     "codesniffer"                
#>  [89] "coffeescript"                "commonlisp"                 
#>  [91] "composer"                    "compressed"                 
#>  [93] "compressedarchive"           "compression"                
#>  [95] "concrete5"                   "coq"                        
#>  [97] "cordova"                     "craftcms"                   
#>  [99] "crashlytics"                 "crbasic"                    
#> [101] "crossbar"                    "crystal"                    
#> [103] "csharp"                      "cuda"                       
#> [105] "cvs"                         "d"                          
#> [107] "dart"                        "darteditor"                 
#> [109] "data"                        "database"                   
#> [111] "datarecovery"                "dbeaver"                    
#> [113] "defold"                      "delphi"                     
#> [115] "dframe"                      "diff"                       
#> [117] "direnv"                      "diskimage"                  
#> [119] "django"                      "dm"                         
#> [121] "docfx"                       "docpress"                   
#> [123] "docz"                        "dotenv"                     
#> [125] "dotfilessh"                  "dotnetcore"                 
#> [127] "dotsettings"                 "dreamweaver"                
#> [129] "dropbox"                     "drupal"                     
#> [131] "drupal7"                     "drupal8"                    
#> [133] "e2studio"                    "eagle"                      
#> [135] "easybook"                    "eclipse"                    
#> [137] "eiffelstudio"                "elasticbeanstalk"           
#> [139] "elisp"                       "elixir"                     
#> [141] "elm"                         "emacs"                      
#> [143] "ember"                       "ensime"                     
#> [145] "episerver"                   "erlang"                     
#> [147] "espresso"                    "executable"                 
#> [149] "exercism"                    "expressionengine"           
#> [151] "extjs"                       "fancy"                      
#> [153] "fastlane"                    "finale"                     
#> [155] "firebase"                    "flashbuilder"               
#> [157] "flask"                       "flex"                       
#> [159] "flexbuilder"                 "floobits"                   
#> [161] "flutter"                     "font"                       
#> [163] "fontforge"                   "forcedotcom"                
#> [165] "forgegradle"                 "fortran"                    
#> [167] "freepascal"                  "fsharp"                     
#> [169] "fuelphp"                     "fusetools"                  
#> [171] "games"                       "gcov"                       
#> [173] "genero4gl"                   "geth"                       
#> [175] "ggts"                        "gis"                        
#> [177] "git"                         "gitbook"                    
#> [179] "go"                          "godot"                      
#> [181] "goodsync"                    "gpg"                        
#> [183] "gradle"                      "grails"                     
#> [185] "greenfoot"                   "grunt"                      
#> [187] "gwt"                         "haskell"                    
#> [189] "helm"                        "hexo"                       
#> [191] "hol"                         "homeassistant"              
#> [193] "hsp"                         "hugo"                       
#> [195] "hyperledgercomposer"         "iar"                        
#> [197] "iar_ewarm"                   "iarembeddedworkbench"       
#> [199] "idapro"                      "idris"                      
#> [201] "igorpro"                     "images"                     
#> [203] "infer"                       "inforcms"                   
#> [205] "inforcrm"                    "intellij"                   
#> [207] "intellij+all"                "intellij+iml"               
#> [209] "ionic3"                      "jabref"                     
#> [211] "java"                        "java-web"                   
#> [213] "jboss"                       "jboss-4-2-3-ga"             
#> [215] "jboss-6-x"                   "jboss4"                     
#> [217] "jboss6"                      "jdeveloper"                 
#> [219] "jekyll"                      "jenv"                       
#> [221] "jetbrains"                   "jetbrains+all"              
#> [223] "jetbrains+iml"               "jgiven"                     
#> [225] "jigsaw"                      "jmeter"                     
#> [227] "joe"                         "joomla"                     
#> [229] "jspm"                        "julia"                      
#> [231] "jupyternotebooks"            "justcode"                   
#> [233] "kate"                        "kdevelop4"                  
#> [235] "kdiff3"                      "keil"                       
#> [237] "kentico"                     "kicad"                      
#> [239] "kirby2"                      "kobalt"                     
#> [241] "kohana"                      "komodoedit"                 
#> [243] "konyvisualizer"              "kotlin"                     
#> [245] "labview"                     "labviewnxg"                 
#> [247] "lamp"                        "laravel"                    
#> [249] "latex"                       "lazarus"                    
#> [251] "leiningen"                   "lemonstand"                 
#> [253] "less"                        "liberosoc"                  
#> [255] "librarian-chef"              "libreoffice"                
#> [257] "lilypond"                    "linux"                      
#> [259] "lithium"                     "logtalk"                    
#> [261] "lsspice"                     "ltspice"                    
#> [263] "lua"                         "lyx"                        
#> [265] "m2e"                         "macos"                      
#> [267] "magento"                     "magento1"                   
#> [269] "magento2"                    "magic-xpa"                  
#> [271] "matlab"                      "maven"                      
#> [273] "mavensmate"                  "mdbook"                     
#> [275] "mean"                        "mercurial"                  
#> [277] "mercury"                     "metals"                     
#> [279] "metaprogrammingsystem"       "meteor"                     
#> [281] "meteorjs"                    "microsoftoffice"            
#> [283] "mikroc"                      "moban"                      
#> [285] "modelsim"                    "modx"                       
#> [287] "momentics"                   "monodevelop"                
#> [289] "mplabx"                      "mule"                       
#> [291] "nanoc"                       "nativescript"               
#> [293] "ncrunch"                     "nesc"                       
#> [295] "netbeans"                    "nette"                      
#> [297] "nikola"                      "nim"                        
#> [299] "ninja"                       "node"                       
#> [301] "nodechakratimetraveldebug"   "notepadpp"                  
#> [303] "nuxt"                        "nuxtjs"                     
#> [305] "nwjs"                        "objective-c"                
#> [307] "ocaml"                       "octave"                     
#> [309] "octobercms"                  "opa"                        
#> [311] "opencart"                    "opencv"                     
#> [313] "openfoam"                    "openframeworks"             
#> [315] "openframeworks+visualstudio" "oracleforms"                
#> [317] "orcad"                       "osx"                        
#> [319] "otto"                        "oxideshop"                  
#> [321] "oxygenxmleditor"             "packer"                     
#> [323] "particle"                    "patch"                      
#> [325] "pawn"                        "perl"                       
#> [327] "perl6"                       "ph7cms"                     
#> [329] "phalcon"                     "phoenix"                    
#> [331] "phpcodesniffer"              "phpstorm"                   
#> [333] "phpstorm+all"                "phpstorm+iml"               
#> [335] "phpunit"                     "pico8"                      
#> [337] "pimcore"                     "pimcore4"                   
#> [339] "pimcore5"                    "pinegrow"                   
#> [341] "platformio"                  "playframework"              
#> [343] "plone"                       "polymer"                    
#> [345] "powershell"                  "premake-gmake"              
#> [347] "prepros"                     "prestashop"                 
#> [349] "processing"                  "progressabl"                
#> [351] "psoccreator"                 "puppet"                     
#> [353] "puppet-librarian"            "purebasic"                  
#> [355] "purescript"                  "putty"                      
#> [357] "pvs"                         "pycharm"                    
#> [359] "pycharm+all"                 "pycharm+iml"                
#> [361] "pydev"                       "python"                     
#> [363] "qml"                         "qooxdoo"                    
#> [365] "qt"                          "qtcreator"                  
#> [367] "r"                           "racket"                     
#> [369] "rails"                       "react"                      
#> [371] "reactnative"                 "reasonml"                   
#> [373] "red"                         "redcar"                     
#> [375] "redis"                       "rhodesrhomobile"            
#> [377] "rider"                       "root"                       
#> [379] "ros"                         "ruby"                       
#> [381] "rubymine"                    "rubymine+all"               
#> [383] "rubymine+iml"                "rust"                       
#> [385] "salesforce"                  "salesforcedx"               
#> [387] "sas"                         "sass"                       
#> [389] "sbt"                         "scala"                      
#> [391] "scheme"                      "scons"                      
#> [393] "scrivener"                   "sdcc"                       
#> [395] "seamgen"                     "senchatouch"                
#> [397] "serverless"                  "shopware"                   
#> [399] "silverstripe"                "sketchup"                   
#> [401] "slickedit"                   "smalltalk"                  
#> [403] "snap"                        "snapcraft"                  
#> [405] "solidity"                    "soliditytruffle"            
#> [407] "sonar"                       "sonarqube"                  
#> [409] "sourcepawn"                  "spark"                      
#> [411] "splunk"                      "spreadsheet"                
#> [413] "ssh"                         "standardml"                 
#> [415] "stata"                       "stdlib"                     
#> [417] "stella"                      "stellar"                    
#> [419] "stylus"                      "sublimetext"                
#> [421] "sugarcrm"                    "svn"                        
#> [423] "swift"                       "swiftpackagemanager"        
#> [425] "swiftpm"                     "symfony"                    
#> [427] "symphonycms"                 "synology"                   
#> [429] "synopsysvcs"                 "tags"                       
#> [431] "tarmainstallmate"            "terraform"                  
#> [433] "terragrunt"                  "test"                       
#> [435] "testcomplete"                "testinfra"                  
#> [437] "tex"                         "text"                       
#> [439] "textmate"                    "textpattern"                
#> [441] "theos-tweak"                 "thinkphp"                   
#> [443] "tla+"                        "tortoisegit"                
#> [445] "tower"                       "turbogears2"                
#> [447] "twincat"                     "typings"                    
#> [449] "typo3"                       "typo3-composer"             
#> [451] "umbraco"                     "unity"                      
#> [453] "unrealengine"                "vaadin"                     
#> [455] "vagrant"                     "valgrind"                   
#> [457] "vapor"                       "venv"                       
#> [459] "vertx"                       "video"                      
#> [461] "vim"                         "virtualenv"                 
#> [463] "virtuoso"                    "visualstudio"               
#> [465] "visualstudiocode"            "vivado"                     
#> [467] "vlab"                        "vue"                        
#> [469] "vuejs"                       "vvvv"                       
#> [471] "waf"                         "wakanda"                    
#> [473] "web"                         "webmethods"                 
#> [475] "webstorm"                    "webstorm+all"               
#> [477] "webstorm+iml"                "werckercli"                 
#> [479] "windows"                     "wintersmith"                
#> [481] "wordpress"                   "wyam"                       
#> [483] "xamarinstudio"               "xcode"                      
#> [485] "xcodeinjection"              "xilinx"                     
#> [487] "xilinxise"                   "xilinxvivado"               
#> [489] "xill"                        "xojo"                       
#> [491] "xtext"                       "y86"                        
#> [493] "yeoman"                      "yii"                        
#> [495] "yii2"                        "zendframework"              
#> [497] "zephir"                      "zig"                        
#> [499] "zsh"                         "zukencr8000"